refactor(analytics): move shared utilities from sysadmin/ to analytics/community/
Phase G の最初の足場として、`analytics/` ドメインを新設し、その配下
`analytics/community/` に共有 utility ファイル群を物理移動する。
中身の変更は無く、`git mv` によるファイル位置変更と import path
書き換えのみ。

  src/application/domain/sysadmin/{bounds,classifiers,aggregations,pagination}.ts
    → src/application/domain/analytics/community/

これらは sysadmin domain でのみ使われる community-level の純関数 +
定数 (member 分類 / 集計 / ページング / 期間境界判定) で、
ユーザ承認済の Plan で「analytics/community/ 配下に集約」する方針を
取った。後続 commit で repository / service / usecase / presenter /
schema を順次移行し、最終的に sysadmin/ ドメインは @deprecated で
旧 query をラップする薄い殻になる。

7 caller (sysadmin/{service,usecase,presenter}.ts + 4 unit tests)
の import path を新ロケーションに更新。動作変更ゼロ、91 unit
tests green。
