+++ src/application/domain/analytics/community/bounds.ts
