+++ src/application/domain/analytics/community/classifiers.ts
