+++ src/application/domain/analytics/community/pagination.ts
